Move Checkbox component to Row/InputTypes
